+++ meringue/templatetags/weboven_base.py
@@ -1,5 +1,8 @@
 # -*- coding: utf-8 -*-
 
+import logging
+
+from django.conf import settings
 from django.template import Library
 from django.utils import timezone
 
@@ -22,3 +25,11 @@ def cop_year():
     year = timezone.now().year
     return year == wn_settings.START_YEAR and year or '%d&mdash;%d' %\
         (wn_settings.START_YEAR, year)
+
+
+@register.simple_tag
+def put_reset():
+    """
+    Шаблонный тег выводит содержимое файла reset.css заключённое в тег style
+    """
+    return '<style>%s</style>' % open(wn_settings.path('meringue/static/css/reset%s.css' % '' if settings.DEBUG else '.min'), 'r').read()